// src/app/core/models/data-response.ts
export class  DataResponse {
    P_NCODE: string;
    P_SMESSAGE: string;
    P_SEMAIL?: any;
    P_SCOD_CLIENT?: any;
    P_SURL_SISTEMA?: any;
    P_NIDCM?: any;
    EListClient: EListClient[];
    EListErrores?: any;
    Data?: any;
  }
  
  export class EListClient {
    P_SCLIENT: string;
    P_NCODE?: any;
    P_SMESSAGE?: any;
    EListClient?: any;
    EClientOperationBE?: any;
    P_NIDDOC_TYPE: string;
    P_SIDDOC: string;
    P_DIG_VERIFICACION?: any;
    P_NUSERCODE: string;
    P_SFIRSTNAME: string;
    P_SLASTNAME: string;
    P_SLASTNAME2: string;
    P_APELLIDO_CASADA?: any;
    P_SSEXCLIEN: string;
    P_NINCAPACITY?: any;
    P_NINCAP_COD?: any;
    P_DBIRTHDAT: string;
    P_DINCAPACITY?: any;
    P_NHEALTH_ORG?: any;
    P_DDEATHDAT?: any;
    P_DWEDD?: any;
    P_SACCOUNT_IN?: any;
    P_NINVOICING?: any;
    P_NSPECIALITY: string;
    P_NCIVILSTA: string;
    P_DAPROBDATE?: any;
    P_SBLOCKADE?: any;
    P_NCLASS?: any;
    P_DDRIVERDAT?: any;
    P_NHEIGHT?: any;
    P_NHOUSE_TYPE?: any;
    P_SLICENSE?: any;
    P_NNOTENUM?: any;
    P_NQ_CARS?: any;
    P_NQ_CHILD?: any;
    P_NRATE?: any;
    P_STAX_CODE?: any;
    P_SSMOKING?: any;
    P_SCUIT?: any;
    P_NTITLE: string;
    P_NOCCUPATION?: any;
    P_NWEIGHT?: any;
    P_SAUTO_CHAR?: any;
    P_SCREDIT_CARD?: any;
    P_NECONOMIC_L?: any;
    P_NEMPL_QUA?: any;
    P_NIMAGENUM?: any;
    P_NAREA?: any;
    P_DDRIVEXPDAT?: any;
    P_NTYPDRIVER?: any;
    P_NDISABILITY?: any;
    P_NLIMITDRIV?: any;
    P_SLEGALNAME?: any;
    P_NAFP?: any;
    P_SBILL_IND?: any;
    P_NNATIONALITY: string;
    P_SDIGIT: string;
    P_DRETIREMENT?: any;
    P_DINDEPENDANT?: any;
    P_DDEPENDANT?: any;
    P_NMAILINGPREF?: any;
    P_NLANGUAGE?: any;
    P_SLEFTHANDED?: any;
    P_SBLOCKLAFT?: any;
    P_SISCLIENT_IND?: any;
    P_SISRENIEC_IND: string;
    P_NCLIENT_SEG?: any;
    P_NCLIENT_SEG_DESCRIP?: any;
    P_SPOLIZA_ELECT_IND?: any;
    P_TI_DOC_SUSTENT?: any;
    P_NU_DOC_SUSTENT?: any;
    P_COD_UBIG_DEP_NAC?: any;
    P_COD_UBIG_PROV_NAC?: any;
    P_COD_UBIG_DIST_NAC?: any;
    P_DEPARTAMENTO_NACIMIENTO?: any;
    P_PROVINCIA_NACIMIENTO?: any;
    P_DISTRITO_NACIMIENTO?: any;
    P_NOMBRE_PADRE?: any;
    P_NOMBRE_MADRE?: any;
    P_FECHA_INSC?: any;
    P_FECHA_EXPEDICION?: any;
    P_CONSTANCIA_VOTACION?: any;
    P_RESTRICCION?: any;
    P_DCOMPDATE: string;
    P_NPERSON_TYP: string;
    P_SNOMUSUARIO: string;
    P_COD_CUSPP?: any;
    P_SPROTEG_DATOS_IND: string;
    P_SGRADO_INSTRUCCION?: any;
    P_FOTO_RENIEC?: any;
    P_FIRMA_RENIEC?: any;
    P_ORIGEN_DATA: string;
    P_SSISTEMA?: any;
    P_SISSEACSA_IND: string;
    P_SISCLIENT_GBD?: any;
    P_SISCLIENT_INTERME: string;
    P_SISCLIENT_CONTRA: string;
    P_SISCORREDOR_IND?: any;
    P_NCANT_TRABAJADORES?: any;
    P_NCANT_USUARIOS?: any;
    P_SCOD_SBS?: any;
    P_SISCONSE_CLAUS_IND?: any;
    P_NID_SECTOR_EMPRESA?: any;
    P_NID_AREA_RESPO?: any;
    P_DFECHA_RECEP_FACTU?: any;
    P_SVENTACONSORCIO_IND: string;
    P_SPARTIDA_ELECTRONICA?: any;
    P_SNAME_COMERCIAL?: any;
    P_DFE_ANIVERSARIO?: any;
    P_SLIDER_CONSORCIO?: any;
    P_SISCLIENT_CRITICO: string;
    P_SBAJAMAIL_IND: string;
    P_SREGIST?: any;
    P_SFOTO?: any;
    P_SFIRMA?: any;
    P_SCONDDOMICILIO?: any;
    P_STIPOCONTR?: any;
    P_SESTADOCONTR?: any;
    EListAddresClient: any[];
    EListPhoneClient: any[];
    EListEmailClient: any[];
    EListContactClient: any[];
    EListCIIUClient: any[];
    EListHistoryClient: any[];
    ElistInfoBancariaClient: any[];
    ElistDocumentosClient: any[];
  }
  
  